ggml : remove ggml-ext.h
